Finish streamlining clueos build scripts

+++ scripts/lc-builds/blueos_clang_omptarget.sh
@@ -7,7 +7,22 @@
 # SPDX-License-Identifier: (BSD-3-Clause)
 ###############################################################################
 
-BUILD_SUFFIX=lc_blueos-clang-upstream-2019.08.15_omptarget
+if [ "$1" == "" ]; then
+  echo
+  echo "You must pass a compiler version number to script. For example,"
+  echo "    blueos_clang_omptarget.sh 10.0.1-gcc-8.3.1"
+  echo "  - or -
+  echo "    blueos_clang_omptarget.sh ibm-10.0.1-gcc-8.3.1"
+  exit
+fi
+
+COMP_VER=$1
+
+BUILD_SUFFIX=lc_blueos-clang-${COMP_VER}_omptarget
+
+echo
+echo "Creating build directory ${BUILD_SUFFIX} and generating configuration in it"
+echo
 
 rm -rf build_${BUILD_SUFFIX} 2>/dev/null
 mkdir build_${BUILD_SUFFIX} && cd build_${BUILD_SUFFIX}
@@ -16,12 +31,12 @@ module load cmake/3.14.5
 
 cmake \
   -DCMAKE_BUILD_TYPE=Release \
-  -DCMAKE_CXX_COMPILER=/usr/tce/packages/clang/clang-upstream-2019.08.15/bin/clang++ \
+  -DCMAKE_CXX_COMPILER=/usr/tce/packages/clang/clang-${COMP_VER}/bin/clang++ \
   -C ../host-configs/lc-builds/blueos/clang_X.cmake \
-  -DENABLE_CUDA=Off \
   -DENABLE_OPENMP=On \
+  -DENABLE_CUDA=Off \
   -DENABLE_TARGET_OPENMP=On \
   -DOpenMP_CXX_FLAGS="-fopenmp;-fopenmp-targets=nvptx64-nvidia-cuda" \
   -DCMAKE_INSTALL_PREFIX=../install_${BUILD_SUFFIX} \
   "$@" \
-  ..
+  ..
